pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dhee"
version = "5.0.0"
description = "Cognition layer for AI agents — persistent memory, performance tracking, and insight synthesis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Sankhya AI Labs"}
]
keywords = ["memory-layer", "cognition", "mcp", "self-evolving", "hyperagent", "ai", "agents", "plugin", "llm", "edge"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
# LLM / embedding providers (install what you use)
gemini = ["google-genai>=1.0.0"]
openai = ["openai>=1.0.0"]
ollama = ["ollama>=0.4.0"]
# Internal providers (not documented, still functional)
nvidia = ["openai>=1.0.0"]
zvec = ["zvec>=0.2.1"]
sqlite_vec = ["sqlite-vec>=0.1.1"]
# Local Qwen stack (CPU-native, zero API cost)
local = ["llama-cpp-python>=0.3", "sentence-transformers>=3.0"]
# Integrations
mcp = ["mcp>=1.0.0"]
api = ["fastapi>=0.100.0", "uvicorn>=0.20.0"]
bus = ["engram-bus>=0.1.0"]
benchmarks = ["huggingface_hub>=0.24.0"]
# Edge/hardware deployment (offline, ONNX embedder)
edge = ["onnxruntime>=1.16"]
# Training (QLoRA fine-tuning)
training = ["unsloth", "datasets>=2.0", "trl>=0.7", "peft>=0.6"]
all = [
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "ollama>=0.4.0",
    "mcp>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "engram-bus>=0.1.0",
    "huggingface_hub>=0.24.0",
    "llama-cpp-python>=0.3",
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "openai>=1.0.0",
    "huggingface_hub>=0.24.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]

[project.scripts]
dhee = "dhee.cli:main"
dhee-mcp = "dhee.mcp_slim:run"
dhee-mcp-full = "dhee.mcp_server:run"
dhee-debugger-api = "dhee.debugger_api:run"

[project.urls]
Homepage = "https://github.com/Sankhya-AI/Dhee"
Repository = "https://github.com/Sankhya-AI/Dhee"
Issues = "https://github.com/Sankhya-AI/Dhee/issues"
Documentation = "https://github.com/Sankhya-AI/Dhee#readme"
Changelog = "https://github.com/Sankhya-AI/Dhee/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["dhee*", "engram*"]
